Add Jump Buffer, Double Jump, & Coyote Time
Replace jump from direct input to a jump buffer to prevent unregistered input
when there is a delay when the input was pressed and the player touches ground.
Add coyote time mechanic that will register the player as still grounded when
the player walks off a ledge for a split second so the player can still jump.
Add jump amount to allow multiple jump, the amount will be decreased for each
time the player jump.