Add 2 enemy type & base death mechanic
Add a moving enemy that will move on the ground and turn around when reaching
a wall or ledge.
Add a ranged enemy that will shoot projectile when player is in range.
Destroy player when colliding with hazard/enemy, further handle of what happen
after death has not been created.